Add support for trimmed tail arrays
Trimmed arrays have a "small" and "large" limit sizes. If the number of
valid elements is <= "small", the array only contains the valid
elements. Otherwise the array is padded to "large" with undefined
values. This scheme could be useful since statically sized arrays can be
copied much faster, but we don't want to always copy lots of padding.